Issue 75: Extract WopiControllerTestBase from WopiControllerTest.
